\name{biostacs-package}
\alias{biostacs-package}
\alias{biostacs}
\docType{package}
\title{
BioStaCs Algorithm Collection
}
\description{
GAS, SVM, GLMMET, Penalized Method. This is a collection and modification.
}
\details{
\tabular{ll}{
Package: \tab biostacs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-17\cr
License: \tab GPL 3.0\cr
}
}
\author{
Yifan Yang

Maintainer:  <yifan.yang@uky.edu>
}
\references{
http://sweb.uky.edu/~yya234/wordpress/?p=75 
}

\keyword{ package }

\examples{
##not run
##not run
}
